+++ lib/pry/commands/easter_eggs.rb
@@ -83,15 +83,10 @@ class Pry
         \____/   \________________________|
       OUTPUT
 
-      move_up =
-        if Helpers::Platform.windows_ansi?
-          proc { |n| "\e[#{n}F" }
-        else
-          proc { |n| "\e[#{n}A\e[0G" }
-        end
+      move_up = "\e[F"
 
       output.puts "\n" * 6
-      output.puts picture.lines.map(&:chomp).reverse.join(move_up[1])
+      output.puts picture.lines.map(&:chomp).reverse.join(move_up)
       output.puts "\n" * 6
       output.puts "** ENV['TERM'] is #{ENV['TERM']} **\n\n"