+++ spec/integration/readline_spec.rb
@@ -3,13 +3,12 @@
 # These specs ensure that Pry doesn't require readline until the first time a
 # REPL is started.
 
-require "shellwords"
 require 'rbconfig'
 
-RSpec.describe "Readline" do
+RSpec.describe "Readline", slow: true do
   before :all do
-    @ruby = RbConfig.ruby.shellescape
-    @pry_dir = File.expand_path(File.join(__FILE__, '../../../lib')).shellescape
+    @ruby = RbConfig.ruby
+    @pry_dir = File.expand_path(File.join(__FILE__, '../../../lib'))
   end
 
   it "is not loaded on requiring 'pry'" do
@@ -17,7 +16,8 @@ RSpec.describe "Readline" do
       require "pry"
       p defined?(Readline)
     RUBY
-    expect(`#{@ruby} -I #{@pry_dir} -e '#{code}'`).to eq("nil\n")
+    out = IO.popen([@ruby, '-I', @pry_dir, '-e', code, err: [:child, :out]], &:read)
+    expect(out).to eq("nil\n")
   end
 
   it "is loaded on invoking 'pry'" do
@@ -26,7 +26,8 @@ RSpec.describe "Readline" do
       Pry.start self, input: StringIO.new("exit-all"), output: StringIO.new
       puts defined?(Readline)
     RUBY
-    expect(`#{@ruby} -I #{@pry_dir} -e '#{code}'`.end_with?("constant\n")).to eq(true)
+    out = IO.popen([@ruby, '-I', @pry_dir, '-e', code, err: [:child, :out]], &:read)
+    expect(out).to end_with("constant\n")
   end
 
   it "is not loaded on invoking 'pry' if Pry.input is set" do
@@ -36,6 +37,7 @@ RSpec.describe "Readline" do
       Pry.start self, output: StringIO.new
       p defined?(Readline)
     RUBY
-    expect(`#{@ruby} -I #{@pry_dir} -e '#{code}'`.end_with?("nil\n")).to eq(true)
+    out = IO.popen([@ruby, '-I', @pry_dir, '-e', code, err: [:child, :out]], &:read)
+    expect(out).to end_with("nil\n")
   end
 end